# docs/openapi.yml
openapi: 3.1.0
info:
  title: Reminder API
  description: Apple Reminders 클론 백엔드 API.
  version: 0.0.1-SNAPSHOT
  contact:
    name: bong
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버

tags:
  - name: ReminderList
    description: 미리 알림 리스트 (사이드바의 사용자 정의 리스트)

paths:
  /api/v1/lists:
    get:
      tags: [ReminderList]
      summary: 리스트 전체 조회
      description: sortOrder 오름차순으로 모든 리스트를 반환한다.
      operationId: findAllReminderLists
      responses:
        '200':
          description: 리스트 배열
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderListResponse'

    post:
      tags: [ReminderList]
      summary: 리스트 생성
      operationId: createReminderList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListCreateRequest'
      responses:
        '201':
          description: 생성된 리스트
          headers:
            Location:
              description: 새로 생성된 리소스 URI
              schema:
                type: string
                example: /api/v1/lists/10
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/lists/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: 리스트 식별자
        schema:
          type: integer
          format: int64

    patch:
      tags: [ReminderList]
      summary: 리스트 부분 수정
      description: 본문에 포함된(null 이 아닌) 필드만 적용한다. 누락된 필드는 변경하지 않는다.
      operationId: updateReminderList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListUpdateRequest'
      responses:
        '200':
          description: 수정된 리스트
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [ReminderList]
      summary: 리스트 삭제
      operationId: deleteReminderList
      responses:
        '204':
          description: 삭제 성공
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    ReminderListCreateRequest:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: 리스트 이름. 공백 불가.
          example: 쇼핑
        color:
          type: string
          minLength: 7
          maxLength: 7
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: 리스트 색상. '#RRGGBB' 형식.
          example: '#FF9500'
        sortOrder:
          type: integer
          format: int32
          minimum: 0
          default: 0
          description: 사이드바 정렬 순서.
          example: 0

    ReminderListUpdateRequest:
      type: object
      description: 부분 업데이트 페이로드. 모든 필드는 선택적.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
          example: 장보기
        color:
          type: string
          minLength: 7
          maxLength: 7
          pattern: '^#[0-9A-Fa-f]{6}$'
          nullable: true
          example: '#0A84FF'
        sortOrder:
          type: integer
          format: int32
          minimum: 0
          nullable: true
          example: 3

    ReminderListResponse:
      type: object
      required: [id, name, color, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: 여행
        color:
          type: string
          example: '#0A84FF'
        sortOrder:
          type: integer
          format: int32
          example: 2
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 UTC.
          example: '2026-04-29T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: ISO-8601 UTC.
          example: '2026-04-29T10:00:00Z'

    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 기계가 식별할 수 있는 오류 코드.
          enum:
            - VALIDATION_ERROR
            - INVALID_ARGUMENT
            - REMINDER_LIST_NOT_FOUND
        message:
          type: string
          description: 사용자에게 노출할 한글 메시지.
        fieldErrors:
          type: array
          description: 필드 단위 검증 오류 (VALIDATION_ERROR 일 때만 채워짐).
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: name
        message:
          type: string
          example: 리스트 이름은 비어 있을 수 없습니다.

  responses:
    BadRequest:
      description: 요청 값이 유효하지 않거나 도메인 검증 실패.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              summary: validation 실패 (필드 단위)
              value:
                code: VALIDATION_ERROR
                message: 요청 값이 유효하지 않습니다.
                fieldErrors:
                  - field: name
                    message: 리스트 이름은 비어 있을 수 없습니다.
            invalidArgument:
              summary: 도메인 검증 실패
              value:
                code: INVALID_ARGUMENT
                message: 색상은 #HEX 형식이어야 합니다.
                fieldErrors: []

    NotFound:
      description: 리소스를 찾을 수 없음.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: REMINDER_LIST_NOT_FOUND
            message: 리스트를 찾을 수 없습니다.
            fieldErrors: []